Refactors all the code into MoreOOP
